Extract component to render legal links
