+++ app/components/layout/footer_component.rb
@@ -1,12 +1,6 @@
 class Layout::FooterComponent < ApplicationComponent
-  delegate :content_block, to: :helpers
-
   def render?
     !Rails.application.multitenancy_management_mode?
-  end
-
-  def footer_legal_content_block
-    content_block("footer_legal")
   end
 
   private
