+++ app/components/layout/legal_links_component.html.erb
@@ -1,0 +1,13 @@
+<ul class="legal">
+  <li><%= link_to t("layouts.footer.privacy"), page_path("privacy") %></li>
+  <li><%= link_to t("layouts.footer.conditions"), page_path("conditions") %></li>
+  <li><%= link_to t("layouts.footer.accessibility"), page_path("accessibility") %></li>
+  <% if feature?(:cookies_consent) %>
+    <li class="cookies-consent-management-footer-content">
+      <a href="#cookies_consent_management" data-open="cookies_consent_management">
+        <%= t("layouts.footer.cookies_consent_management") %>
+      </a>
+    </li>
+  <% end %>
+  <%= raw footer_legal_content_block %>
+</ul>
