+++ app/components/layout/legal_links_component.rb
@@ -1,0 +1,7 @@
+class Layout::LegalLinksComponent < ApplicationComponent
+  delegate :content_block, to: :helpers
+
+  def footer_legal_content_block
+    content_block("footer_legal")
+  end
+end
